+++ comicfn2dict/comicfn2dict.py
@@ -1,3 +1,0 @@
-"""API import source."""
-from comicfn2dict.parse import comicfn2dict  # noqa: F401
-from comicfn2dict.unparse import dict2comicfn  # noqa: F401